scrapped template-to-auto
Was attempting to address #536 but alas we must retain use of explicit templating in some utility functions. This is because they themselves call other templated functions, and in my testing, the compiler was not able to deduce that only the type of some arguments were consulted by the nested invocations, and so threw an error that `constexpr` expressions were not constant.

I think it's not worthwhile to mix explicit templating with `auto`, so we should retain all `template` unless we can solve above